requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pypdf
python-dotenv
flask
